Refactor waterfall files
